firmware: axis: Moved tracking_rates to separated header file
